make dungeon rings green if you can get some items.
previously, if you could go into a dungeon and get some but not all of its chests,
it had a yellow ring. this is not consistent with the use of yellow on overworld
chests. i changed it to green. this involves new availabilities "partial"
and "glitchpartial" where they used to be using "possible" and "glitchpossible"

it's easy to change the color from green to some unique other color if desired